edit controllers as per apidocs, all tests fixed
